+++ test_project/settings.py
@@ -32,6 +32,7 @@ INSTALLED_APPS = [
     'eav',
 ]
 
+
 MIDDLEWARE = [
     'django.middleware.security.SecurityMiddleware',
     'django.contrib.sessions.middleware.SessionMiddleware',
@@ -69,7 +70,9 @@ DATABASES = {
     },
 }
 
+
 DEFAULT_AUTO_FIELD = 'django.db.models.AutoField'
+EAV2_PRIMARY_KEY_FIELD = 'django.db.models.AutoField'
 
 
 # Password validation